+++ primal-memory/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Rename `ByteArray` -> `PrimArray`
 * Improve `fromList` conversion
+* Get rid of `allocByteCountMem` in favor of more general `allocMem`
 
 ## 0.1.0